backend/requirements.txt
# FastAPI Backend Dependencies - Production
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Google APIs (for Gmail integration)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# File Processing
pandas==2.1.4
openpyxl==3.1.2
aiofiles==23.2.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Background Jobs
celery==5.3.4  # Will be replaced with ThreadPoolExecutor in YAGNI approach

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Logging
structlog==23.2.0
rich==13.7.0

# Development Tools
python-dotenv==1.0.0